Major revision
TopicSim class, deleted run_simulation and run_mitigation, added config file
