+++ run.sh
@@ -1,5 +1,5 @@
 cd 4_simulation/src
-python3 run_simulation.py
+python3 run_mitigation.py
 cd ../../5_analysis/src
 python3 process_data.py
 python3 make_plots.py
